+++ src/frontend/js/pages/TeacherDashboardContractsLayout/index.ts
@@ -1,0 +1,2 @@
+export { TeacherDashboardCourseContractsLayout } from './TeacherDashboardCourseContractsLayout';
+export { TeacherDashboardOrganizationContractsLayout } from './TeacherDashboardOrganizationContractsLayout';
