+++ src/frontend/js/pages/TeacherDashboardContractsLayout/styles.scss
