--- src/frontend/js/pages/TeacherDashboardContractsLoader/index.ts
@@ -1,2 +1,0 @@
-export { TeacherDashboardCourseContractsLoader } from './TeacherDashboardCourseContractsLoader';
-export { TeacherDashboardOrganizationContractsLoader } from './TeacherDashboardOrganizationContractsLoader';
