+++ src/frontend/scss/components/_index.scss
@@ -21,7 +21,7 @@
 @import '../../js/pages/DashboardCreditCardsManagement/styles';
 @import '../../js/pages/DashboardOrderLayout/styles';
 @import '../../js/pages/DashboardPreferences/styles';
-@import '../../js/pages/TeacherDashboardContractsLoader/styles';
+@import '../../js/pages/TeacherDashboardContractsLayout/styles';
 @import '../../js/pages/TeacherDashboardCourseLoader/CourseRunList/styles';
 @import '../../js/pages/TeacherDashboardTraining/styles.scss';
 @import '../../js/widgets/Dashboard/styles';
